requirements.txt
flask>=3.0.0
pyyaml>=6.0.1
click>=8.1.7
virtualenv>=20.0.0
